Figured out what was wrong with my character encoding conversions.
Turned out my assumptions that a single codepoint is max 4 utf8 bytes and that composing and such doesn't change that was right after all. I simply misinterpreted the result of the conversion functions. The source was indeed exhausted, but that was because it was already reading the bytes of the next character, not the first.