Drastically changed PylirDialect.
There now is single object type called Pylir_Object which represents a Python object. It has a single mutable parameter which is an optional FlatSymbolReference pointing to the actual type object if known. As it is mutable it does not partake in object equality.

The other builtin types are value types which are not objects. They have to be boxed to be turned into a Pylir_Object. This allowed a few things, including removing Pylir_Float and Pylir_Bool and their ops in favour of using the standard dialect and builtin types