Replace py.getAttr and py.setAttr with py.getSlot and py.setSlot.
The former were meant to access the dictionary of the object, if it had one. That is quite the loaded operation however. Instead I figured I should be implementing slots instead and make it serve as the atom operation of all the attribute access for primitives. The `__dict__` attribute will also just be a slot in an object after all.