Who would have thunk, I did have a bug.
The "Flaky" test from last commit now works perfectly on Windows too.

The issue was the compiler generated code not allocating enough bytes for objects which lead to writes past an allocation. Fixed the error by simplifying the whole thing and storing the size of the (object - slots) / sizeof(void*), instead of an offset in it. Also added an UnhandeledExceptionFilter to MSVC so that exceptions thrown by libunwind, that are not caught, lead to resumption of the code. That way the `pylir_raise` function code can handle the uncaught exception. MinGW already has that in their startup files, no need there.