Fix Undefined Behaviour in py.str.equal
We use memcmp to compare two strings once they have been determined to be of equal size. If that size is 0 however then it'd be called with nullptrs which is Undefined Behaviour. We insert a guard that checks whether the size is 0, and if it is directly concludes that the strings are equal