Mark now broken tests as XFAIL
This is bit unfortunate but I have no clue what to do for now. I am having an increasingly hard time testing the codegen output of the frontend and I am concerned this might be an indication that my IR is not high level enough.

For now I will probably try to get more features into the builtins so that I can do more and more execution tests