Fix issue with dangling block argument in SSA Builder.
This may occur if a block argument of an open block has been deemed trivial and therefore replaced with a new value but the block argument is still the definition in successor blocks. Those are not replaced.

I implemented a VERY naive solution by simply iterating over the definitions map and manually replacing them. This will have to do for now.

The LLVM way of getting around it would be via the `TrackingVH` which is part of a `llvm::Value`s use list and therefore always up to date with the replaced values.
We don't have such a thing in MLIR however and might never given the implementation complexity