Proof of concept stackmap
Implemented a custom stackmap that is easily readable by the runtime which contains instruction on how to read all local variables in the current frame of any suspended call (that isn't a gc leaf function).
To place the statepoints and find the minimum lifeset I imported the RewriteStatepoints4GC pass from LLVM but modified it for now to use the deopt arguments of statepoints to get pure spill semantics.
As soon as I want to use a relocating GC I can simply turn a switch as well.