Turn all previously quasi subtypes of PyObject into actual subtypes of PyObject while retaining the standard layout.
Standard layout types are allowed to have subtypes as long as only either the base or the derived type contains non-static members. To retain that property, PyObject was split in two: The PyObjectStorage which is the common object header of all subclasses, and PyObject which has ALL the methods of PyObject that subclasses can now use. Implicit conversion from subclasses to PyObject also removes many redundant casts and the implicit conversion operators.