Rename `py.linearContains` op to `py.tuple.contains` op and lower it to LLVM instead of expanding it.
Expanding it would only be of value if the expanded implementation would be then again consist of high level components that could itself be optimized, but not the Op as a whole. This is the case of eg. the expansions in `py.makeTuple` and friends, but I feel is not the case for `py.tuple.contains`.
Whether this is the case for `py.mroLookup` is TBD, but I believe so as well (it consist of smaller individual ops that could be optimized yes, but it as a whole could be optimized at least as effectively I believe)