[Parser] Reimplement the whole AST
This can actually be called an AST now and has less of a resemblance of an APT. Improvements are via the introduction of a AbstractIntrusiveVariant we have a more convenient sum type interface than with std::variant and second, by not representing the syntactic elements of the input source in the tree, but rather the semantic elements. Eg. we don't have an 'expression_list' anymore, but instead a `TupleConstruct` which may be produced by other grammar constructs as well.