[CodeGen] Introduce new builtins for all operators
We now have the concept of the `BuiltinsModule` which is pythons `builtins` module, as well as the compiler builtins, which are functions defined and used by the compiler. All stages of the compiler will assume that all these builtins are able to be used.
The next step will be trying to convert the python builtins module to actual python code + intrinsics