[CodeGen] Generate cells in a deterministic order
This was previously non-deterministic as it was dependent on the iteration order of an unordered_set. There are multiple approaches to fixing this, but I opted for what I felt like was the cleanest solution without source breaks or the like which was to just use a std::set instead.