[CodeGen] Allow writing python `builtins` module in python
This works via the use of intrinsics, which for the moment bridge the gaps of either things simply not yet implemented in the compiler, or for things that are not possible to express in python.

It is designed to be an incremental process, allowing a class or function be migrated one at a time with the eventual goal of eliminating the manually generated within BuiltinsImpl.cpp to be eliminated entirely.

Writing python code to implement python should allow for a vastly better development experience and also serves as an excellent test for the compiler frontend and optimizer