[FoldHandles] Handle single store case for references as well
This previously only handled constant objects by promoting the constant to a global value and referencing that. In the case that the single store already refers to a global value, it is not necessary to do so and instead all loads can be replaced to a reference to that global value.