[Monomorph] Properly cap size of recursive types and handle external function calls
These are two odd cases to be put into the same patch but they each lead to the same code path so here we are.
Since one could, via recursion, just keep on creating new infinite length types we have to cap it in some way as to not make the compiler run forever and take infinite amounts of memory.