[Monomorph] Disable handling of recursion immediately
The method implemented has been faulty from the very beginning and rewriting of most of the pass will be required to properly handle it. The gist of the issue is that we are handling dependencies on the function level, aka. that functions depend on each other and wait on fully evaluation to continue. This however, creates circular dependencies that in reality do not exist. The proper solution would be instead to track dependencies on the operation level, making call results dependent on the values passed into a return. This would create similar constructs to a normal loop which can be run until fixpoint