[Diag] Improve DiagnosticsBuilder API through the use of flag arguments and better defaults.
A lot of the optional arguments may not be used in which case many `std::nullopt` appeared in the code previously. This patch fixes that by making these arguments flags instead, which can appear in any order and do not have a designated position.

Additionally, smarter defaults are now in use to reduce the amount of times one does have to pass arguments. The colour palette to use is now automatically deduced from the current message a label is being added to, and the primary colour is used by default unless specified otherwise.