[Diag] Add DiagnosticsManager and rework how diagnostics are emitted in the various components of the compiler
Diagnostics are now emitted not by bubbling up a string or similar but via callbacks from a diagnostic managers.
This has multiple advantages including:
* Unified path for errors, warnings and notes in all components and even for location-less environmental errors
* The parser may now do error recovery and is able to emit more than one error

Only the minimal amount of refactoring has been done to make it compile (which still amounts to a lot), but in particular the parsers error handling has not yet changed. This patch just makes it possible as it makes it possible to detach the error path from the actual parsing. As a conservative first step, the control flow has not been changed, meaning it still mostly immediately exist upon any errors. This will be changed in future patches.