[Main] Add `-verify` implementation to check diagnostics output
This allows the use of 'expected' comments, as can be seen in 'mlir-opt's '-verify-diagnostic' or clangs '-verify' (which is more capable however), to verify that only expected diagnostics of the given kind, message on the specific line are emitted.
Current error message tests in lexer and parser are currently implemented as unit tests. With this patch it'll be possible to make them proper integration tests executed via 'pylir'.

Notes on the implementation: The regex matching for finding the 'expected' comments currently uses the CRTP library. The reason for this is that Documents currently operate on 'char32_t' as well as its very convenient API. This is not ideal for the used regex specifically however, as the maximum match is not bounded. CRTP uses stack memory in such cases, hence running into the rist of stack overflow.
The binary size increase is also not great. In the case that documents are made utf-8 in the future, we should probably use `llvm::Regex`.