[Parser] Improve target handling
We were previously too lenient when parsing some target constructs. This patch now properly errors out when discovering multiple iterable unpacking in a target. Additionally, the target list for a `del` statement is not allowed to have any iterable unpacking at all. This is not properly documented in the python reference sadly, but can be seen here:
https://github.com/python/cpython/blob/4e4d35d3325fb1e05dad43bd1ec73f14c3c5dc0a/Grammar/python.gram#L592