[Main] Only add workaround CLI option for AArch64 triples
This is technically not correct, as we should only be adding this option if AArch64 is linked in.
Otherwise, we get an error that it is an unknown command line option.
This is the next best option however, as we shouldn't get thi far without having the target enabled anyways.