[Main] Change approach from last commit
This workaround currently only exists for *.ll files which have a target triple set to AArch64 (for development). Problem is that the target triple from the LLVM file does not affect the toolchain (and I am not sure it should), so the check does not reliably work.

As a workaround for now, check if AArch64 was compiled with LLVM. We always link in every target so this should be fine