[CI] Further speed up rebuilds of LLVM from cache by using optimized TableGen instances
This further reduces LLVM build times by a minute or two. Reason is that without this patch, TableGen instances compiled with sanitizers are run to regenerate the auto generated headers as they cannot be cached by ccache. These are quite a bit slower than if we use a release build of TableGen.