[CI] Add Windows MSVC CI
This Windows configuration currently uses MSVC with debug STL to build LLVM and then uses clang-cl to build Pylir, also with the debug STL. 

Using MSVC to build Pylir is currently not supported, but I am using it to build LLVM due to a defieceny in ccache, making it not capapble of caching clang-cl compilations at the moment. 

Config using clang with MinGW and libc++ and sanitizers will be added in the future.