[Parser] Fix parsing of lambdas which required turning the parser stateless
There is an ambiguity in the grammar of lambdas due to parameters of lambdas being able to be typed, preventing us from previously parsing them correctly.

A consequence of this is that in particular edges cases, AST nodes have to be moved around. Problem is that parsing is currently stateful, in the sense that it has side effects as namespace resolution is currently being performed during parsing of identifiers. This has now become problematic as the parsed content actually switches scopes when being moved to resolve lambdas ambiguous grammar.

All semantic checking and namespace resolution was hence extracted out of the parser and refactored into its own pass using a classic visitor pattern.