[PylirPy] Change python integer to arith dialect conversion to exclusively make use of index
Except i1, we exclusively make use of index instead of the other integer types. Not having to support all the other types makes the code easier as we don't need to handle the other integer types cases.

More importantly, there are now two ways of turning an index into a python integer: `py.int.fromUnsigned` interprets the index as an unsigned integer, while `py.int.fromSigned` interprets it as signed. It is therefore now possible to create a negative integer from an index value, which was previously not possible.