[CI] Add GCC as Compiler on Linux
This now also tests that LLVM and Pylir can be built with GCC-10. No sanitizers are used with GCC-10 due to not being properly supported in LLVM. Instead we now have both have GCC-10 and Clang builders working without sanitizers which run the tests much faster.

The reason GCC-10 is used is because it's the first GCC version to support lld.