[PylirPy] Add explicit `$hash` parameters to `dict` operations
The point of these is that it forces the calculations of the hashes to be done in the frontend aka by python code, instead of doing it from C++. Doing so from C++ would force us to go through the whole unoptimized dynamic lookup paths, while doing so from python code allows the `hash` call to be optimized with the might of the whole optimization pipeline.

This patch does not yet actually make use of this however. It is currently just the required frontend and dialect changes. Backend changes to LLVM and to the runtime API will be in a follow up.