[PylirPy] Change `py.mroLookup` result type to a potentially unbound object, instead of the separate success boolean
Having a separate result for information that is essentially encoded in the first result made it redundant.

Additionally, it has caused some uglyness when used as an intrinsic in builtins.py. Since it had multiple results we returned it as a tuple of its results. Problem with that however, is that now we are potentially creating a tuple with an unbound value. I have not yet thought about whether that should be allowed or not, but with this patch we don't need to worry about it for now :)