[HandleLoadStoreElim] Simplify clobber handling by replacing block args dependent on clobbers immediately
If during SSA building of the replacements of loads and stores of handles, a block argument is created which depends on a clobber, we are not allowed to use that block argument to replace the load, as the re-load is required if coming from a branch with a clobber.

Previously this was done by doing a DFS of the use list of the special clobber value and finding all block arguments that are dependent on it. The new approach uses a new callback in SSABuilder to optimize block arguments. It is used to replace a block argument that has at least one clobber value with the clobber value instead. This leads to all block arguments depending on clobbers to be destroyed immediately and less load instructions being delayed for replacement and instead immediately resolved as required.