[SSABuilder] Add marking algorithm as optimization to SSABuilder
The gist of what it does is delay the creation of block arguments in the recursive lookup case until it is absolutely necessary. This may either be due to being in a loop and hence needing to break the cycle, or if block argument is actually required to join two unrelated values. This leads to a massive reduction of back and forth of creating block arguments and immediately deleting them after.