[Main] Rework output file name logic
If the input is of a known filetype with a known file extension, replace the output file extension with a suitable one for the requested output kind.
Make sure to always create the default output filenames relative to the working directory.

Also properly handle output to special files like /dev/null. This sadly can't be tested in lit AFAIK