[tblgen] Add new backend for use with Op generation that can generate interface methods via OpVariableDecorators
This allows us close to the exact same syntax as one is used to with SideEffects for MLIR. The tblgen interface is generic and works by generating the body by splitting it into 3 parts after each other:
* Prologue which is defined by the `ExtendableInterfaceMethods`
* Body which is the body defined by the `OpVariableGen` which is pasted and substituted for every argument annotated
* Epilogue which is also defined by the `ExtendableInterfaceMethods`

The only difference to MLIRs SideEffects is that we do sadly still need to manually add the interface trait to the list.

This has been implemented with CaptureInterface.td as a proof of concept.