[PylirPy] Add new interfaces `CopyObjectInterface` and `OnlyReadsValueInterface` and use them in fold operations.
These interfaces are for operations that do a builtin type copy as well as operations that only read the value part of an immutable object respectively. When the immutable value parts of a result of a copy operation are read we can then fold it to instead read the input of the copy operation, with the hopes of eliminating the copy operation entirely and allowing further optimizations on the read.