[TrialInliner] Improve trial inliner by running it in bottom-up order of the callgraph
This is done by walking over the SCCs of the callgraph in reverse topological order. This has the advantage that inlining a direct call, the callee that is being inlined has already been optimized. As a side effect it saves on runtime as the function body of the callee is also not optimized multiple times.

That said, due to the extremely high numbers of indirect calls (we are compiling Python after all), it just yields rather mild improvements in the resulting IR. The SCCs are formed prior to any inlining occurring and does not take into account new direct calls that are formed during inlining.