[SORA] Make the SROA generic by operating purely over interfaces
This commit introduces two interface within SROAInterfaces.td which are used for allocation functions and read/write ops respectively. The main motivation of doing so is making the code a lot cleaner, decoupled as well as allowing future reuse of the interfaces' implementations in the operation. As a side effect, SROA is now completely generic and does not depend on any dialects.