[MemorySSA] Partition uses and defs by resource
Since reads and writes of one resource cannot overlap with another resource it is always correct partition use def chains by resource. This is then applied to the PylirPy dialect, which currently differentiates between handle, object, list and dict resource. A Write to the list part of an object therefore no longer interferes with the use-def chains of slot accesses on the same object.

This could technically also be achieved through improving alias analysis' ModRef behaviour but would require building up alias anylsis infrastructure that goes beyond the very basic one in MLIR.
Additionally, we'd requick block arguments in MemorySSA to be optimized for equal results, which they currently aren't.