[PylirPy] Add MemoryFoldUsage implementation to `py.load`
With the new MemorySSA changes it is now viable and possible for stores and loads to handles to be optimized as well. The HandleLoadStoreElimination pass is technically speaking a simple specialization of MemorySSA, simplified for the more trivial special case of simple load and store. In the future if load forwarding is just as capable, and dead store elimination is implemented, it can be removed in favour of using those.