[CI] Update ubuntu version to 22.04 and msys2 repos
Reason for the ubuntu version is mostly for newer tools, especially newer versions of ccache.
Updating msys2 each build also makes more sense as that is how users generally use it.
The ubuntu update causes builds to switch from using clang 10 to clang 11 now, which is the lowest clang version in the 22.04 repos.