[PylirPy] Relax restriction on being able to use `RefAttr` within verifiers.
Through a combination of reimplementing the `pylir-opt` driver, as well as upstream MLIR allowing for textual or bytecode to be imported without verifying, we can now run the finalize pass before running any verifiers.
The effect of this is that we can now once again share code between the verifier and the normal compiler implementation.
Those changes will be done in follow-up commits however.