[Main] Add initial support for Darwin toolchains
The Darwin toolchain contains the logic for driving the linking process on Darwin based platforms and is the final step to convert the object files created by the compiler into actual executables.
The cruft of this patch is emitting the linker call to ld64.lld and finding the installed SDK and Clang installation.

This is the final missing piece to make Pylir compile on and for MacOS on x86_64.
More architectures (arm64) should be mostly a matter of ABI support in generated LLVM.