[CI] Add macOS 12 CI runner
The CI runner will ensure that Pylir keeps on working on macOS and will also run macOS specific tests for us.

Changes that had to be done boil down to disabling lld, since that is not available on macOS by default, and working around an odd issue with ccache, where AppleClangs mtime kept on changing each job.