Test out new caching strategy
The LLVM build output directory gets cached directly and not even builds have to happen if in cache.
To reduce the size of the build directory, thing archives are also used.
This should hopefully lead to enough size reduction and better compression that the resulting cache files are less than 0.5 and 1.5 GB of the current respective ccache instances.

ccache will remain for Pylir however due to high frequent source changes, unlike the only occasional version bump for LLVM.