[SROA] Fix invalid materializing of undef logic
The logic used did not work in the general case as a SSABuilder may call the callback to produce an undefined value at the end of a block, if sealing, as well.
The current aggregate may at that point not be the same as the one that originally created the block argument.

While this is far from an ideal solution, it is a valid stop gap that decouples SROA from the PylirPy dialect for the time being.