[SROA] Add new resource parameter to read write callbacks that acts as an additional scope/argument for identifying values within an aggregate
Just the attributes is error-prone in the sense that special care has to be taken that two unrelated values within the aggregate do not end up using the same identifying key. This becomes somewhat difficult since the various kinds of keys ops will want to use often boil down to simple integer or strings.

To work around this issue the resource parameter was added as an additional argument to identifying a value.
Using a resource has the nice symmetry of being able to reuse the meaning and definitions of the various resources that are already used in memory effects of the ops participating in SROA.