[PylirPy] Switch `py.getSlot` and `py.setSlot` to use integer indexing instead of string lookup
The later requires doing a linear search through the slot tuple of the metatype which is currently generated late in the backend and fed into LLVM, giving us no chance to optimize the code in our higher level dialects.
Additionally, it was not possible to use runtime values to select the slots.
The new approach can take a runtime index to select the slot that should be accessed. In quite a lot of cases, especially in the builtins code and frontend generated setSlot operations, the slot to be accessed is also constant/known ahead of time and can then simply be hardcoded. Using enums and global constants/intrinsics, the code in builtins and CodeGen ends up also with less "magic strings" and more strong typing