[Inliner] Add new heuristic based inliner, repacling the trial inliner
This inliner has been in the works for a while and is finally good enough for what I'd consider an initial version. In tests, it consistently creates smaller modules than the old inliner while having a faster runtime as well.

This inliner is essentially an odd mix of GCCs and LLVMs approach. The inliner is a module level pass that collects all callsites in the module, rates their cost, and then uses a min-heap queue to inline all callsites, starting with ones of the lowest cost. After a call has been inlined, the new callsites are added to the queue and any callers of the callsites containing callable, are reevaluated and their cost updated.

The inline cost estimation values is taken from LLVM, where most instruction has the abstract unit cost 5 with the exceptions of things like calls and more. To estimate the callsites inline cost, the size of all operations of the callee are summed up. While doing so however, constant folding is done, using known constants from the callsite. If these succeed, the operation is regarded as free. If through constant folding any paths become unreachable, their containing ops become free as well.

There is almost certainly room for a lot more improvements in the future. A non comprehensive list of things to implement or test out are:
* Estimate SROA optimizations after inlining and subtract it from the cost
* Test out optimizing one SCC at a time instead of the module. This would have the advantage of having a topological ordering, leading to optimized callees being inlined into callers. This could also be combined with a function-level optimization pass pipeline, instead of module level as it currently is, to interleave more optimizations between inlinings. LLVM does this.

As a side node, I had to disable the monomorph pass in the pipeline as it had miscompiles and also does not cope well with being executed more than once in the compilation pipeline.