[Inliner] Remove inlining history in favour of penalizing repeated inlining through a call-site
This is achieved not through a single inline history, but one per call edge. This is more appropriate for our module level inliner than the old one as we do not have a bottom up order like LLVMs inliner does.
We are also keeping the history across inlining iterations by attaching a `edge_id` attributes to the call-sites.
This is admittedly fragile, but for the time being I don't see a better alternative until MLIR has listeners or another concept similar to LLVMs ValueHandle.
The new history is also more memory intensive which is mitigated by reusing dead indices. It could potentially be improved however.