[Main] Add support for building and linking the runtime with sanitizers
This is mostly meant as a developer option but also has the advantage of simplifying the cmake logic to build the runtime as well. It no longer requires compiling its dependencies twice, once with and once without sanitizers.
The way this was accomplished is by teaching 'pylir' how to link ASAN, TSAN and UBSAN.

For the time being this is only supported on Windows with MinGW and on Linux. MacOS support should not be difficult to implement in the future (but we'd want a CI runner for that as well).

All tests that require linking an executable also automatically add the required sanitizer flags to be able to link and execute the programs.

Note that Pylir generated code does not yet generate ASAN calls.