[Main] Add support for sanitizer codegen
This allows for Pylir generated code to also be instrumented by the given sanitizers. The use-cases for this are rather simple: ASAN in combination with a ASAN compiled runtime library should aid a lot in debugging the GC and catching codegen issues causing memory errors and TSAN may even be useful for end users in the future (but is still just a developer option for now).

It should be noted that in the current commit, the codegen is opt-in via a command line flag.
The reason is very simple: the sanitizer codegen is tightly coupled wiht the sanitizer runtime and technically speaking, both have to have the exact same version. Opting-out guarantees that the current configurations, including CI keep working while still being able to use the feature as developers if ones host compilers version happens to work.