[PylirHIR] Initial introduction of `pyHIR` dialect
This commits adds a new high level python MLIR dialect meant to more closely represent python programs as they appear in source.
This is meant to achieve multiple goals:
* Move complexity out of the frontend into more intuitive and easier to maintain MLIR lowering passes. Friction in creating ops in the dialect from the frontend should generally be considered a red flag.
* Better allow high level optimizations. Things that would be considered trivially constant folded in python source code currently create too much code in the `py` dialect as an example, but also inlining or optimization of nonlocal variables should be easier.
* Generally easier to analyse for the sake of optimization

This commit only adds initial top level ops and the dialect as a whole should be considered experimental. Next steps will be to create a `CodeGenNew` module to test out lowering from AST to the new dialect and to evolve it accordingly.