[Py] Add `py.func` and `py.return` ops
Motivation for doing so is to be able to deviate from the semantics of the `func` dialect. In particular, this is one of the steps required to making `!py.dynamic` a parameterized type to make type analysis state an essential part of the IR.
As an added benefit we can also write our own more suited patterns for function conversion and remove the `py` prefix from most operations (This could be achieved without creating a custom op as well though).