[SSABuilder] Fix use-after-free when optimizing block arguments
A trivial block argument that was originally recorded as dependent on a just-replaced block argument might be replaced and deleted as a side effect of other block argument simplifications. The simple fix is to just use a value tracker to always have the up-to-date replacement of the block argument. It is sadly not too trivial without hacks whether a replacement occurred, but trying to simplify the replacement is essentially a noop anyways and does not impact correctness or optimality.