[test] add support for running individual lit tests
To allow lit to run individual tests it has to know about both the normal cfg file and the site.cfg files for a specific test. We achieve this the same way LLVM does, by creating `pylir-lit.py` at configure time in cmake and setting a config mapping pointing at the configured lit.site.cfg.py.

This way it is now possible to run `pylir-lit.py path/to/test.py` successfully, where `pylir-lit.py` is found in `bin` of the CMake build directory