[PylirPy] Replace use of `.` as "namespace" in dialect mnemonic with `_`
The main advantage is that it resolves a parsing ambiguity. We use `py` as default dialect within `py.func` to allow omission of `py.` as prefix. MLIR refuses to do that for any ops that have a `.` in their name however, as a `int.add` op may either be parsed as `py.int.add` or `add` of the dialect `int`.

We simply circumvent this issue altogether by using `_` instead of `.` from now on.