[Py][PyHIR][PyMem] Make use of the new MLIR properties for inherent attributes
The new properties are great in that they are saved inline with the op and accessed like a struct field would, instead of the dictionary lookup as previously.

There is more room for future improvements in other dialects as well, as properties also does not require one to pack data into attributes in the first place. This is likely to be useful for the MemorySSA dialect but is not part of this commit.