[CI] Split LLVM build into its own workflow
Splitting it into its own workflow allows us to run the workflow independently of any other workflows.
This first and foremost enables us to run the LLVM builds weekly to keep them in cache or manually

It also allows us to better reuse the cache between workflows. The clang-tidy runner should no longer need to build there own version of LLVM.
This is sadly not yet easily possible for the GitHub pages runner