[Py] Remove current `Monomorph` implementation (#35)
The current implementation is unused in the pipeline due to both
correctness and speed issues. The most problematic thing about it
however, is that it requires a lot of infrastructure which has
essentially been built just for the pass.

With the broader goal of restructure Pylir to focus on lowering and
designing the IR first and foremost, these parts of the infrastructure
lead to slowing down development velocity heavily.