[PylirToLLVM] Remove incorrect `noalias` attribute on `pylir_gc_alloc`
The `noalias` attribute on return types of functions has additional special semantics in LLVM. According to the LangRef, it makes LLVM assume that the pointer returned by the function is unique AND that it comes from the system allocator directly.

The latter has bad consequences in our case however. It makes LLVM assume that stores to the objects memory that are not read are dead and can be removed. This is incorrect in our case however as the GC relies on code setting the type object and attempts reading it. LLVM deleting the store of the type object however leads to it remaining null, and the GC crashing when trying to parse the heap during garbage collection.

As an alternative, LLVM supports the `allockind` attribute which informs it that the function comes from a system allocator. It is currently not nicely modelled in MLIRs LLVM Dialect however and left for a future PR