[Py] Remove `TypeRefinableInterface`, extra types in dialect and supporting infrastructure
This interface was mostly used by `TypeFlow` only. There were some other users of the interface, specifically in folding, but these did not make use of the power afford by the interface. These could therefore be heavily simplified.
This is part of focus on lowering first and removing tech debt in Pylir in the hopes of increasing velocity