[docs] Treat warnings as errors (#37)
This way if the docs were to emit any errors, the CI should fail instead
of silently creating issues.

Note that in the C++ world, treating warnings as errors is somewhat of
an anti-pattern. This is because newer compiler versions often introduce
new warnings, breaking the use of newer toolchains. The hope is that
this is not as problematic in Sphinx due to 1) the sphinx version being
locked in the requirements file and 2) there not being a lot of warnings
and tendency for false-positives or opinionated code