[test] Remove `-O3` from execution tests (#40)
This was originally added as I found it "useful", to essentially stress
test the optimizer pipeline. This remains true, but is the wrong
approach. The -O3 heavily impacts testing times and generates
"test-coverage" from the wrong place. Rather, the tests in Execution
test 1) the implementation in `src/python` and 2) integration with the
compiler.

Arguably this should be split in the future. For now remove `-O3` to
heavily reduce test time. Future changes enabled by this PR can then:
- add a nightly in the pipeline that does use -O3 to stress test it
without impacting iteration time and PR merge time
- compile these python files with `pylir` as part of the cmake build
process

Tiny note: a bug in the GC was found and is fixed here as well. Our
runtime testing infrastructure is currently too lacking to easily add a
regression test however...