[Py] Remove `TypeRefinableInterface` and supporting infrastructure (#39)
This interface was mostly used by `TypeFlow` only. There were some other
users of the interface, specifically in folding, but these did not make
use of the power afforded by the interface. These were therefore
heavily simplified. This is part of focus on lowering first and removing
tech debt in Pylir in the hopes of increasing velocity