[docs] Fix stateful build
Building the docs would often create errors due to files from previous builds remaining in the created source directory.
This commit fixes that by always creating the source directory anew.

Ideally we'd not have to create a source directory for the docs using cmake but use the in-repo sources. This is currently impossible however as part of our sources are auto-generated and sphinx only supports a single root directory for its sources.