[Py] Introduce `#py.globalValueAttr` (#47)
This is the very first step in trying to get rid of `#py.ref` and the
`py.globalValue` operation.
The new attribute based version has the advantages of not requiring the
"op-ref" split, automatically performing DCE via reachability, and being
able to be both referenced and created at any point in time. It most
importantly also cleans up a lot of code and will make it possible to
get rid of `#py.ref` in the future.