[cmake] Add warning if found LLVM has a different revision than the pinned
A version mismatch is a comparatively hard to find issue. Ideally, it'd simply be a build break but in worse cases it might even be a runtime error. Either way, thinking of checking the LLVM version and then manually comparing the revisions of both required and installed is a hassle.
This commit therefore adds some logic to CMake to verify the revision found with the revision required, displaying a clear warning if mismatched.